LayoutTests/fast/dom/Range/range-intersectsNode-exception-expected.txt

Test for Bug: 75799 - Calling intersectsNode on a detached range should throw

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS range.intersectsNode(document.getElementById('b1')) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS successfullyParsed is true

TEST COMPLETE
